requirements.txt
alembic==0.8.3
beautifulsoup4==4.4.1
Flask==0.10.1
Flask-SQLAlchemy==2.1
gunicorn==19.3.0
igo-python==0.9.3
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.3
MarkupSafe==0.23
mecab-python3==0.7
psycopg2==2.6.1
pymongo==3.2.2
python-editor==0.4
SQLAlchemy==1.0.3
Werkzeug==0.10.4
